backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
opencv-python==4.8.1.78
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
SQLAlchemy==2.0.23
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
psutil==5.9.6
flake8==6.1.0
mypy==1.7.1
bandit==1.7.5

# AI Integration Dependencies (Phase 2)
# VEO API SDK (Google Video Generation)
google-cloud-aiplatform==1.38.1
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Weather API Integration
python-weather==1.1.0
requests==2.31.0

# Task Scheduling
celery==5.3.4
redis==5.0.1
APScheduler==3.10.4

# AI Utilities
numpy==2.3.3
pandas==2.1.3
pydantic==2.5.2
scikit-learn==1.7.2